Add README with permissions and groups info

+++ advanced_features_and_security/LibraryProject/bookshelf/views.py
@@ -1,39 +1,51 @@
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required, permission_required
-from .models import Book
-from .forms import BookForm
+from .models import Book  # Make sure you have a Book model
+from .forms import BookForm  # Optional: if you have a form for creating/editing books
 
+# View to list all books
 @login_required
 @permission_required('bookshelf.can_view', raise_exception=True)
-def list_books(request):
+def book_list(request):
     books = Book.objects.all()
-    return render(request, 'bookshelf/list_books.html', {'books': books})
+    return render(request, 'bookshelf/book_list.html', {'books': books})
 
+
+# View to create a new book
 @login_required
 @permission_required('bookshelf.can_create', raise_exception=True)
-def add_book(request):
+def book_create(request):
     if request.method == 'POST':
         form = BookForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('list_books')
+            return redirect('book_list')
     else:
         form = BookForm()
-    return render(request, 'bookshelf/add_book.html', {'form': form})
+    return render(request, 'bookshelf/book_form.html', {'form': form})
 
+
+# View to edit an existing book
 @login_required
 @permission_required('bookshelf.can_edit', raise_exception=True)
-def edit_book(request, pk):
+def book_edit(request, pk):
     book = get_object_or_404(Book, pk=pk)
-    form = BookForm(request.POST or None, instance=book)
-    if form.is_valid():
-        form.save()
-        return redirect('list_books')
-    return render(request, 'bookshelf/edit_book.html', {'form': form})
+    if request.method == 'POST':
+        form = BookForm(request.POST, instance=book)
+        if form.is_valid():
+            form.save()
+            return redirect('book_list')
+    else:
+        form = BookForm(instance=book)
+    return render(request, 'bookshelf/book_form.html', {'form': form})
 
+
+# View to delete a book
 @login_required
 @permission_required('bookshelf.can_delete', raise_exception=True)
-def delete_book(request, pk):
+def book_delete(request, pk):
     book = get_object_or_404(Book, pk=pk)
-    book.delete()
-    return redirect('list_books')
+    if request.method == 'POST':
+        book.delete()
+        return redirect('book_list')
+    return render(request, 'bookshelf/book_confirm_delete.html', {'book': book})
